tried to accelerate

+++ scripts/helpers.py
@@ -5,11 +5,11 @@ import numpy as np
 
 
 def get_empty_vectors(vectors_number: int) -> np.ndarray:
-    return np.zeros((vectors_number, 3), dtype=np.float)
+    return np.zeros((vectors_number, 3), dtype=np.float32)
 
 
 def get_empty_float_scalars(scalars_number: int) -> np.ndarray:
-    return np.zeros(scalars_number, dtype=np.float)
+    return np.zeros(scalars_number, dtype=np.float32)
 
 
 def get_empty_int_scalars(scalars_number: int) -> np.ndarray:
